refactor(governance): 拆分第二批历史近阈值热路径文件
